extend key from uint32 2 int64

+++ api.go
@@ -74,7 +74,7 @@ type API interface {
 	// value 插入数据对象
 	//
 	// 返回 hashKey
-	PutD(key string, value interface{}) (uint32, error)
+	PutD(key string, value interface{}) (int64, error)
 	// SetD 设置数据，如果存在将被覆盖，如果不存在，则新建
 	//
 	// 向_default表中新增或更新一条数据，key相同则覆盖
@@ -84,7 +84,7 @@ type API interface {
 	// value 插入数据对象
 	//
 	// 返回 hashKey
-	SetD(key string, value interface{}) (uint32, error)
+	SetD(key string, value interface{}) (int64, error)
 	// GetD 获取数据
 	//
 	// 向_default表中查询一条数据并返回
@@ -104,7 +104,7 @@ type API interface {
 	// value 插入数据对象
 	//
 	// 返回 hashKey
-	Put(databaseName, formName, key string, value interface{}) (uint32, error)
+	Put(databaseName, formName, key string, value interface{}) (int64, error)
 	// Set 设置数据，如果存在将被覆盖，如果不存在，则新建
 	//
 	// 向指定表中新增或更新一条数据，key相同则覆盖
@@ -118,7 +118,7 @@ type API interface {
 	// value 插入数据对象
 	//
 	// 返回 hashKey
-	Set(databaseName, formName, key string, value interface{}) (uint32, error)
+	Set(databaseName, formName, key string, value interface{}) (int64, error)
 	// Get 获取数据
 	//
 	// 向指定表中查询一条数据并返回
@@ -138,7 +138,7 @@ type API interface {
 	// formName 表名
 	//
 	// value 插入数据对象
-	Insert(databaseName, formName string, value interface{}) (uint32, error)
+	Insert(databaseName, formName string, value interface{}) (int64, error)
 	// Update 更新数据
 	//
 	// 向指定表中新增或更新一条数据，key相同则覆盖
@@ -212,7 +212,7 @@ type Database interface {
 	// 返回 hashKey
 	//
 	// update 本次是否执行更新操作
-	put(formName string, key string, value interface{}, update bool) (uint32, error)
+	put(formName string, key string, value interface{}, update bool) (int64, error)
 	// Get 获取数据
 	//
 	// 向_default表中查询一条数据并返回
@@ -230,7 +230,7 @@ type Database interface {
 	// value 插入数据对象
 	//
 	// 返回 hashKey
-	insert(formName string, value interface{}, update bool) (uint32, error)
+	insert(formName string, value interface{}, update bool) (int64, error)
 	// querySelector 根据条件检索
 	//
 	// formName 表名
@@ -246,7 +246,7 @@ type Database interface {
 // 提供表基本操作方法
 type Form interface {
 	WriteLocker
-	getAutoID() *uint32           // getAutoID 返回表当前自增ID值
+	getAutoID() *int64            // getAutoID 返回表当前自增ID值
 	getID() string                // getID 返回表唯一ID
 	getName() string              // getName 返回表名称
 	getDatabase() Database        // getDatabase 返回数据库对象
@@ -284,13 +284,13 @@ type Index interface {
 	// value 存储对象
 	//
 	// update 本次是否执行更新操作
-	put(originalKey string, key uint32, update bool) IndexBack
+	put(originalKey string, key int64, update bool) IndexBack
 	// get 获取数据，返回存储对象
 	//
 	// originalKey 真实key，必须string类型
 	//
 	// key 索引key，可通过hash转换string生成
-	get(originalKey string, key uint32) (interface{}, error)
+	get(originalKey string, key int64) (interface{}, error)
 	// recover 重置索引数据
 	recover() error
 }
@@ -310,7 +310,7 @@ type Nodal interface {
 	// value 存储对象
 	//
 	// update 本次是否执行更新操作
-	put(key string, hashKey, flexibleKey uint32, update bool) IndexBack
+	put(key string, hashKey, flexibleKey int64, update bool) IndexBack
 	// get 获取数据，返回存储对象
 	//
 	// key 真实key，必须string类型
@@ -318,7 +318,7 @@ type Nodal interface {
 	// hashKey 索引key，可通过hash转换string生成
 	//
 	// flexibleKey 下一级最左最小树所对应真实key
-	get(key string, hashKey, flexibleKey uint32) (interface{}, error)
+	get(key string, hashKey, flexibleKey int64) (interface{}, error)
 	getDegreeIndex() uint8 // getDegreeIndex 获取节点所在树中度集合中的数组下标
 	getPreNode() Nodal     // getPreNode 获取父节点对象
 }
@@ -342,7 +342,7 @@ type Link interface {
 	getSeekStart() uint32         // value最终存储在文件中的起始位置
 	getSeekLast() int             // value最终存储在文件中的持续长度
 	getValue() interface{}
-	put(key string, hashKey uint32) *indexBack
+	put(key string, hashKey int64) *indexBack
 	get() (interface{}, error)
 }
 
@@ -351,7 +351,7 @@ type IndexBack interface {
 	getLocker() WriteLocker       // 索引文件所对应level2层级度节点
 	getLink() Link                // 索引对应节点对象子集
 	getKey() string               // 索引对应字符串key
-	getHashKey() uint32           // put hash keyStructure
+	getHashKey() int64            // put hash keyStructure
 	getErr() error
 }
